gateway/pyproject.toml
[project]
name = "django-prefect-gateway"
version = "0.1.0"
description = "FastAPI gateway for Prefect integration"
requires-python = "==3.13.*"

dependencies = [
    "fastapi==0.128.6",
    "uvicorn[standard]==0.40.0",
    "prefect==3.6.16",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "pyjwt==2.11.0",
    "python-multipart==0.0.22",
    "slowapi==0.1.9",
    "redis==7.1.1",
]

[project.optional-dependencies]
dev = []

test = [
    "pytest==8.4.2",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
]

[tool.setuptools.packages.find]
include = ["api", "core", "middleware"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "asyncio: Async tests",
]
addopts = [
    "--verbose",
    "-p", "no:warnings",  # Suppress warnings for cleaner output
]
asyncio_mode = "auto"  # Enable automatic async test detection